move some tests to different files
and fix an issue in the Mangopay callback tests
